Modulo detector id's to fit into <0;n) range. Add new builder error & some plumbing.

+++ src/game/level/builder/mod.rs
@@ -86,8 +86,8 @@ struct IntermediateCycleData {
 	pub linked_cycle: IntermediateLinkStatus,
 	/// Outgoing oneways
 	pub outgoing_one_way_links: Vec<OneWayIntermediateData>,
-	/// Detectors, pairs of detector ID's and their offsets on this cycle.
-	pub placed_detectors: Vec<(usize, i32)>,
+	/// Detectors, pairs of detector ID's and their offsets on this cycle. (The second, positional, index is the index of the vertex this detector comes *after*)
+	pub placed_detectors: Vec<(usize, usize)>,
 }
 
 #[derive(Clone, Debug)]
